# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Service Assurance AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /dashboard/recent-tickets:
    get:
      operationId: getRecentTickets
      tags: [dashboard]
      summary: Get recent open tickets
      parameters:
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Recent tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketWithRelations"
...
  /customers:
    get:
      operationId: getCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerWithRelations"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerRequest"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/ai-test:
    post:
      operationId: aiTest
      tags: [ai]
      summary: Test AI status normalization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiTestRequest"
      responses:
        "200":
          description: AI test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiNormalizeResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, ops, customer]
        customerId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
        - updatedAt
...
    Customer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        accountNumber:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        primaryContactName:
          type: string
          nullable: true
        primaryContactEmail:
          type: string
          nullable: true
        primaryContactPhone:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - createdAt
        - updatedAt
    CustomerWithRelations:
      allOf:
        - $ref: "#/components/schemas/Customer"
        - type: object
          properties:
            sites:
              type: array
              items:
                $ref: "#/components/schemas/Site"
            services:
              type: array
              items:
                $ref: "#/components/schemas/Service"
            tickets:
              type: array
              items:
                $ref: "#/components/schemas/Ticket"
    CreateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        accountNumber:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        primaryContactName:
          type: string
          nullable: true
        primaryContactEmail:
          type: string
          nullable: true
        primaryContactPhone:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
        - status
    UpdateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        accountNumber:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        primaryContactName:
          type: string
          nullable: true
        primaryContactEmail:
          type: string
          nullable: true
        primaryContactPhone:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    Site:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        siteName:
          type: string
        address1:
          type: string
          nullable: true
        address2:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        siteCode:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - siteName
        - createdAt
        - updatedAt
...
    Service:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        siteId:
          type: string
        vendorName:
          type: string
        serviceType:
          type: string
          enum: [DIA, Broadband, SD-WAN, Voice, Wireless, Other]
        circuitId:
          type: string
          nullable: true
        bandwidth:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, pending, down, impaired, disconnected]
        installDate:
          type: string
          nullable: true
        monthlyRecurringCharge:
          type: number
          nullable: true
        supportReference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - siteId
        - vendorName
        - serviceType
        - status
        - createdAt
        - updatedAt
...
    Ticket:
      type: object
      properties:
        id:
          type: string
        ticketNumber:
          type: string
        customerId:
          type: string
        siteId:
          type: string
          nullable: true
        serviceId:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        source:
          type: string
          enum: [manual, email, api]
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [new, investigating, vendor_engaged, dispatch_scheduled, monitoring, resolved, closed]
        outageType:
          type: string
          enum: [outage, impairment, informational, unknown]
        vendorTicketId:
          type: string
          nullable: true
        assignedToUserId:
          type: string
          nullable: true
        openedAt:
          type: string
          format: date-time
        lastUpdatedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        nextEscalationAt:
          type: string
          format: date-time
          nullable: true
        slaTargetMinutes:
          type: number
          nullable: true
        aiSummary:
          type: string
          nullable: true
        aiNormalizedStatus:
          type: string
          nullable: true
        aiCustomerUpdate:
          type: string
          nullable: true
        aiLastGeneratedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - ticketNumber
        - customerId
        - title
        - source
        - severity
        - status
        - outageType
        - openedAt
        - lastUpdatedAt
        - createdAt
        - updatedAt
    TicketWithRelations:
      allOf:
        - $ref: "#/components/schemas/Ticket"
        - type: object
          properties:
            customer:
              $ref: "#/components/schemas/Customer"
            site:
              $ref: "#/components/schemas/Site"
            service:
              $ref: "#/components/schemas/Service"
            assignedTo:
              $ref: "#/components/schemas/User"
...
    AiNormalizeResponse:
      type: object
      properties:
        normalizedStatus:
          type: string
        ticket:
          $ref: "#/components/schemas/Ticket"
      required:
        - normalizedStatus
        - ticket
...
    AiTestRequest:
      type: object
      properties:
        text:
          type: string
      required:
        - text
...