# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Service Assurance AI API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceServiceType } from "./serviceServiceType";
import type { ServiceStatus } from "./serviceStatus";

export interface Service {
  id: string;
  customerId: string;
  siteId: string;
  vendorName: string;
  serviceType: ServiceServiceType;
  circuitId?: string | null;
  bandwidth?: string | null;
  status: ServiceStatus;
  installDate?: string | null;
  monthlyRecurringCharge?: number | null;
  supportReference?: string | null;
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}